Feat restarant item

+++ src/main/java/com/example/techchallenge/service/UserService.java
@@ -1,9 +1,9 @@
 package com.example.techchallenge.service;
 
-import com.example.techchallenge.dto.AddressResponse;
-import com.example.techchallenge.dto.UserResponse;
+import com.example.techchallenge.dto.Response.AddressResponse;
+import com.example.techchallenge.dto.Response.UserResponse;
 import com.example.techchallenge.enums.UserRoles;
-import com.example.techchallenge.dto.UserRequest;
+import com.example.techchallenge.dto.Request.UserRequest;
 import com.example.techchallenge.exception.UserNotFoundException;
 import com.example.techchallenge.exception.InvalidCredentialsException;
 import com.example.techchallenge.entities.AddressEntity;
@@ -26,15 +26,18 @@ public class UserService {
     private static final String EMPTY_PASSWORD_MESSAGE = "Senha não pode ser vazia.";
     private static final String EMPTY_NAME_MESSAGE = "Nome não pode ser vazio.";
     private static final String EMAIL_ALREADY_EXISTS_MESSAGE = "Email já está em uso.";
+    private static final String USER_CANNOT_BE_DELETED_OWNER_RESTAURANT = "Usuário não pode ser excluído, pois é dono de um restaurante.";
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     private final AddressService addressService;
+    private final RestaurantService restaurantService;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, AddressService addressService) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, AddressService addressService,RestaurantService restaurantService) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
         this.addressService = addressService;
+        this.restaurantService = restaurantService;
     }
 
     public UserEntity createUser(UserRequest userRequest, String userRoles) {
@@ -66,6 +69,9 @@ public class UserService {
 
     public void delete(Long id) {
         UserEntity entity = getUserById(id);
+        if (restaurantService.getRestaurantByOwnerId(id)) {
+            throw new IllegalArgumentException(USER_CANNOT_BE_DELETED_OWNER_RESTAURANT);
+        }
         userRepository.delete(entity);
     }
 
@@ -140,7 +146,8 @@ public class UserService {
             user.getId(),
             user.getName(),
             user.getEmail(),
-            addressResponse
+            addressResponse,
+            user.getRoles()
         );
     }
